Setup infrastructure and create modules
