+++ src/utils.js
@@ -1,0 +1,5 @@
+const render = (template, parent, place) => {
+  parent.insertAdjacentHTML(place, template);
+};
+
+export {render};
