+++ src/view/day.js
@@ -1,0 +1,3 @@
+const day = `<li class="trip-days__item  day"></li>`;
+
+export {day};
