+++ src/view/event.js
@@ -1,0 +1,39 @@
+const event = `
+  <li class="trip-events__item">
+    <div class="event">
+      <div class="event__type">
+        <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Event type icon">
+      </div>
+      <h3 class="event__title">Sightseeing in Chamonix</h3>
+      <div class="event__schedule">
+        <p class="event__time">
+          <time class="event__start-time" datetime="2019-03-19T11:20">14:20</time>
+          &mdash;
+          <time class="event__end-time" datetime="2019-03-19T13:00">13:00</time>
+        </p>
+        <p class="event__duration">1H 20M</p>
+      </div>
+      <p class="event__price">
+        &euro;&nbsp;<span class="event__price-value">50</span>
+      </p>
+      <h4 class="visually-hidden">Offers:</h4>
+      <ul class="event__selected-offers">
+        <li class="event__offer">
+          <span class="event__offer-title">Book tickets</span>
+          &plus;
+          &euro;&nbsp;<span class="event__offer-price">40</span>
+        </li>
+        <li class="event__offer">
+          <span class="event__offer-title">Lunch in city</span>
+          &plus;
+          &euro;&nbsp;<span class="event__offer-price">30</span>
+          </li>
+      </ul>
+      <button class="event__rollup-btn" type="button">
+        <span class="visually-hidden">Open event</span>
+      </button>
+    </div>
+  </li>
+`;
+
+export {event};
