+++ src/view/events-list.js
@@ -1,0 +1,3 @@
+const eventsList = `<ul class="trip-events__list"></ul>`;
+
+export {eventsList};
