+++ src/view/trip-days.js
@@ -1,0 +1,3 @@
+const tripDays = `<ul class="trip-days"></ul>`;
+
+export {tripDays};
